iba_api/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
setuptools==57.0.0
sqlparse==0.4.2
